refactor(package): fix fragment/activity package
