+++ app/src/androidTest/java/com/example/ezvault/DeleteItemTest.java
@@ -23,7 +23,7 @@ import com.example.ezvault.model.utils.ItemBuilder;
 import com.example.ezvault.model.ItemList;
 import com.example.ezvault.model.User;
 import com.example.ezvault.utils.UserManager;
-import com.example.ezvault.viewmodel.MainActivity;
+import com.example.ezvault.view.MainActivity;
 import com.google.firebase.Timestamp;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
