+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
         </provider>
 
         <activity
-            android:name=".viewmodel.MainActivity"
+            android:name=".view.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
