+++ app/src/main/java/com/example/ezvault/view/NewUserFragment.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.viewmodel;
+package com.example.ezvault.view;
 
 import android.os.Bundle;
 import android.util.Log;
