+++ app/src/main/java/com/example/ezvault/view/SearchFragment.java
@@ -1,4 +1,4 @@
-package com.example.ezvault.viewmodel;
+package com.example.ezvault.view;
 
 import android.os.Bundle;
